Refactor : 피드백반영
